normalize_heatmap
[b]Description:[/b]
This was a simple indicator to indicate the heatmap area of an asset price, in a relative given time period. In default the lookback period was set to 50 bars, indicating the current state of the price within the previous lookback period. The color scheme was using the rainbow palette, which set blue as the cooling-off area, and red as the heating area. The indicator doesn't take into account momentum strategy and thus doesn't consider the future direction of the asset price. Note: cooling-off area, can be considered to entry or adding position as a DCA strategy.

Data Normalize:
norm = (x - min) / (max - min)

[b]Feature: [/b]
[list]
[*] Heatmap color condition
[*] Weighted Moving average (Additional)
[/list]